+++ UnityProject/Assets/Scripts/Managers/GameData.cs
@@ -192,7 +192,7 @@ public class GameData : MonoBehaviour
 		return false;
 	}
 
-	private async void AttemptAutoJoin(Action<String> OnLoginSuccess)
+	private async void AttemptAutoJoin(Action<string> OnLoginSuccess)
 	{
 		await Task.Delay(TimeSpan.FromSeconds(0.1));
 
@@ -202,8 +202,7 @@ public class GameData : MonoBehaviour
 			$"Loading user profile for {FirebaseAuth.DefaultInstance.CurrentUser.DisplayName}");
 
 		await FirebaseAuth.DefaultInstance.CurrentUser.TokenAsync(true).ContinueWithOnMainThread(
-			async task =>
-			{
+			task => {
 				if (task.IsCanceled || task.IsFaulted)
 				{
 					LobbyManager.Instance.lobbyDialogue.LoginError(task.Exception?.Message);
@@ -226,7 +225,7 @@ public class GameData : MonoBehaviour
 
 		GameScreenManager.Instance.serverIP = ip;
 
-		if (!String.IsNullOrEmpty(token))
+		if (string.IsNullOrEmpty(token) == false)
 		{
 			LobbyManager.Instance.lobbyDialogue.ShowLoggingInStatus("Verifying account details..");
 			var refreshToken = new RefreshToken();
@@ -242,7 +241,7 @@ public class GameData : MonoBehaviour
 				return;
 			}
 
-			if (!string.IsNullOrEmpty(response.errorMsg))
+			if (string.IsNullOrEmpty(response.errorMsg) == false)
 			{
 				Logger.LogError($"Something went wrong with hub token validation {response.errorMsg}", Category.DatabaseAPI);
 				LobbyManager.Instance.lobbyDialogue.LoginError($"Could not verify your details {response.errorMsg}");
