+++ UnityProject/Assets/Scripts/Managers/JobAllocator.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using Player;
 
 namespace Managers
 {
