+++ UnityProject/Assets/Scripts/Managers/NetworkManagement/CustomNetworkManager.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using UnityEngine;
@@ -13,6 +12,7 @@ using IgnoranceTransport;
 using Initialisation;
 using Messages.Server;
 using UnityEditor;
+using Util;
 
 public class CustomNetworkManager : NetworkManager, IInitialise
 {
