+++ UnityProject/Assets/Scripts/NPC/AI/MobObjective.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using Doors;
 using UnityEngine;
