+++ UnityProject/Assets/Scripts/Objects/Alien/AlienEggCycle.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections;
 using UnityEngine;
-using Mirror;
 using AddressableReferences;
 using Messages.Server.SoundMessages;
 
@@ -29,10 +28,6 @@ namespace Alien
 		[SerializeField]
 		private EggState initialState = EggState.Growing;
 
-		// TODO This is entirely unused and is creating a compiler warning.
-		[Tooltip("Allows mappers to have eggs that won't start cycle once spawned, but are still intractable")][SerializeField]
-		private bool freezeCycle = false;
-
 		[Tooltip("A reference for the facehugger mob so we can spawn it.")]
 		[SerializeField]
 		private GameObject facehugger = null;
@@ -51,7 +46,6 @@ namespace Alien
 			objectAttributes = GetComponent<ObjectAttributes>();
 		}
 
-
 		public void OnSpawnServer(SpawnInfo info)
 		{
 			incubationTime = UnityEngine.Random.Range(60f, incubationTime);
@@ -60,12 +54,10 @@ namespace Alien
 			registerObject.SetPassable(false, false);
 		}
 
-
 		public void OnDespawnServer(DespawnInfo info)
 		{
 			StopAllCoroutines();
 		}
-
 
 		private void UpdatePhase(EggState state)
 		{
@@ -95,23 +87,20 @@ namespace Alien
 			UpdateExamineMessage();
 		}
 
-
-		IEnumerator GrowEgg()
+		private IEnumerator GrowEgg()
 		{
 			yield return WaitFor.Seconds(incubationTime / 2);
 			UpdatePhase(EggState.Grown);
 		}
 
-
-		IEnumerator WaitForHatchEgg()
+		private IEnumerator WaitForHatchEgg()
 		{
 			yield return WaitFor.Seconds(incubationTime / 2);
 
 			StartCoroutine(HatchEggAnimation());
 		}
 
-
-		IEnumerator HatchEggAnimation()
+		private IEnumerator HatchEggAnimation()
 		{
 			spriteHandler.ChangeSprite(OPENING_SPRITE);
 			yield return WaitFor.Seconds(OPENING_ANIM_TIME);
@@ -119,7 +108,6 @@ namespace Alien
 
 			Spawn.ServerPrefab(facehugger, gameObject.RegisterTile().WorldPositionServer);
 		}
-
 
 		private void UpdateExamineMessage()
 		{
@@ -146,7 +134,6 @@ namespace Alien
 			objectAttributes.ServerSetArticleDescription(examineMessage);
 		}
 
-
 		public void ServerPerformInteraction(HandApply interaction)
 		{
 			switch (currentState)
@@ -172,7 +159,6 @@ namespace Alien
 			}
 		}
 
-
 		private void Open(HandApply interaction)
 		{
 			StopAllCoroutines();
@@ -185,7 +171,6 @@ namespace Alien
 			StartCoroutine(HatchEggAnimation());
 			registerObject.SetPassable(false, true);
 		}
-
 
 		private void Squish(HandApply interaction)
 		{
@@ -203,7 +188,6 @@ namespace Alien
 			UpdatePhase(EggState.Squished);
 			registerObject.SetPassable(false, true);
 		}
-
 
 		private void FeelsSlimy(Interaction interaction)
 		{
