+++ UnityProject/Assets/Scripts/Player/PlayerManager.cs
@@ -4,6 +4,7 @@ using System.Text.RegularExpressions;
 using Newtonsoft.Json;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using Player;
 
 public class PlayerManager : MonoBehaviour
 {
