+++ UnityProject/Assets/Scripts/Systems/Antagonists/Antags/Blob/Blob.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine;
 using Blob;
+using Player;
 
 namespace Antagonists
 {
