+++ UnityProject/Assets/Scripts/Systems/Antagonists/Antags/Blob/BlobPlayer.cs
@@ -7,7 +7,6 @@ using System.Text;
 using Light2D;
 using Mirror;
 using UnityEngine;
-using AddressableReferences;
 using Random = UnityEngine.Random;
 using EpPathFinding.cs;
 using HealthV2;
