+++ UnityProject/Assets/Scripts/Systems/Antagonists/Antags/Cargonian.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
-using Messages.Server;
 using UnityEngine;
+using Messages.Server;
+using Player;
 
 namespace Antagonists
 {
