+++ UnityProject/Assets/Scripts/Systems/Antagonists/Antags/Survivor.cs
@@ -1,5 +1,5 @@
-﻿using System.Collections;
-using UnityEngine;
+﻿using UnityEngine;
+using Player;
 
 namespace Antagonists
 {
