+++ UnityProject/Assets/Scripts/Systems/GameModes/Extended.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using Player;
 
 namespace GameModes
 {
