+++ UnityProject/Assets/Tests/Asset/NetworkManagerTests.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 using UnityEditor;
 using UnityEditor.SceneManagement;
 using UnityEngine;
+using Util;
 
 namespace Tests.Asset
 {
